# film-manager-implementation/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: http://localhost:3001/
...
paths:
...
  "/api/films/public/{filmId}/reviews/{reviewerId}/coreviewer":
    post:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}/coreviewer
      security:
      - cookieAuth: []
      summary: Appoint a coreviewer for a review
      description: A coreviewer, with coreviewer id specified in the request body, is appointed for the review of the film with ID filmId issued to the reviewer with ID reviewerId. The corresponding review is thus edited. This operation can only be performed by the invited reviewer.
      operationId: appointCoreviewer
      parameters:
      - name: filmId
        in: path
        description: ID of the film
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the reviewer
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: The id of the user appointed as co-reviewer
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CoreviewerAssignment"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Review"
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (user is not the invited reviewer)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (a coreviewer has already been appointed for the review)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/{filmId}/reviews/{reviewerId}":
    get:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      summary: Retrieve a review that has been issued/completed for a film
      description: 'The review of the film with ID filmID issued to the user with
        ID reviewerId is retrieved. This operation does not require authentication. '
      operationId: getSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose reviews must be retrieved
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Review"
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      security:
      - cookieAuth: []
      summary: Complete a review
      description: The review of the film with ID filmId and issued to the user with
        ID reviewerId is completed. This operation only allows setting the "completed"
        property to the "true" value, and changing the values of the "reviewDate",
        "rating", and "review" properties. This operation can be performed only by
        the invited reviewer.
      operationId: completeSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose review must be completed
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated Review object
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Review"
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request (the completed property is absent, or the review
            property is absent while no co-reviewer has been appointed)
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (the user is not the reviewer)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
      - api/films/public/{filmId}/reviews/{reviewerId}
      security:
      - cookieAuth: []
      summary: Delete a review invitation
      description: The review of the film with ID filmId and issued to the user with
        ID reviewerId is deleted. This operation can only be performed by the owner,
        and only if the review has not yet been completed by the reviewer.
      operationId: deleteSingleReview
      parameters:
      - name: filmId
        in: path
        description: ID of the film whose review invitation must be deleted
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerId
        in: path
        description: ID of the user to whom the review has been issued
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden (the user is not the owner)
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (the review has already been completed)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
components:
  schemas:
...
    Review:
      description: schema for Review data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
        coreviewerId:
          type: integer
          description: unique identifier of the the user who has received the coreview invitation from a reviewer
        completed:
          type: boolean
          description: true if the review has been completed, false otherwise
          default: 'false'
        reviewDate:
          type: string
          format: date
          description: the date when the review has been completed by the invited
            user
        rating:
          type: integer
          description: the rating the reviewer has given to the film
          minimum: 1
          maximum: 10
        review:
          type: string
          description: a textual description of the review
          maxLength: 1000
        self:
          type: string
          format: uri
        drafts:
          type: string
          format: uri
          description: the url for the drafts of the review
      required:
      - filmId
      - reviewerId
    CoreviewerAssignment:
      description: schema for appointing a co-reviewer to a review
      type: object
      additionalProperties: false
      properties:
        coreviewerId:
          type: integer
          description: unique identifier of the user appointed as co-reviewer
      required:
      - coreviewerId
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid